refactor: make _Filter from game filter public
We can (and will) resuse it in other screens.
